# api/public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: MovieMind API
  version: 1.0.0
  description: |
    Public REST API for movies, people, TV series, and TV shows with AI-powered descriptions.
    Uses slug-based resource identifiers. When data is missing and the corresponding
    feature flag is enabled, endpoints return 202 Accepted and queue generation.
servers:
  - url: http://localhost:8000/api
    description: Local
...
paths:
  /v1/movies:
    get:
      tags: [Movies]
      summary: List movies
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Free-text search (title, director, genre)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
...
  /v1/movies/{slug}/collection:
    get:
      tags: [Movies]
      summary: Get movie collection/franchise
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
    basicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        release_year:
          type: integer
          nullable: true
        director:
          type: string
          nullable: true
...